lawyer_layer/requirements.txt

fastapi
uvicorn
python-multipart
streamlit
google-generativeai
requests
beautifulsoup4
PyPDF2
python-docx
pillow
pytesseract
python-jose
